Fixes: NB#286569 - song title, artist and album name in Ukrainian language are not recognized in the device
RevBy: Holger Schröder

Details:

Improve encoding detection for the Ukrainan legacy encodings
KOI8-U and windows-1251. MCharsetDetector had problems there.
This does *not* fix the example given in bug#286569#c2, but this
example had

    Title  : йіїЄєЦЙУКІВхїЇ
    Album  : йіїЄєЦЙУКІВхїЇ
    Artist: йіїЄєЦЙУКІВхїЇ

in windows-1251 encoding. I.e. just random characters and on top of
them the same in title, album, and artist.  Usually concatenating
title, album, and artist and then detecting increases the chance of a
successful detection but if the same random junk is just repeated 3
times, it does not help.

Although this patch does not fix the detection of the above sample
in bug#286569#c2, it does improve detection of KOI8-U and windows-1251
for longer, more realistic texts.