Fixes: NB#289026 - Regional format; AM/PM (enne keskpäeva/pärast keskpäeva) is clipped in some applications for Estonia.
RevBy: Berthold Krevert

Details:

Use "AM/PM" for Estonia as well instead of translated AM/PM markers.

Estonia uses 24 hour mode as the default, we agreed to use AM/PM for
countries which use 24 hour mode as default if 12 hour mode is forced
because the translations for AM/PM for such countries where 12 hour
mode is unusual are often weird and often also far too long.  We made
very few exceptions for example for Persian because a Latin
(left-to-right) AM/PM marker really does not look nice in date and
time formats which have some parts in Arabic script. And it may even
cause problems with bidi reordering. But unless we have such a strong
reason to use translated AM/PM markers even though 24 hour mode is the
default in that region, we use AM/PM always when 24 hour mode is the
default.

Even in German, the original CLDR data

            AmPmMarkers{
                "vorm.",
                "nachm.",
            }

is too long as well and looks weird even to me as a native German speaker.

For Estonian, CLDR had the data

            AmPmMarkers{
                "enne keskpäeva",
                "pärast keskpäeva",
            }

which is certainly far too long to use anywhere where we want
to use such a marker.