Fixes: NB#288453 - The regional date format for Slovenia is not right.
RevBy: Berthold Krevert

Details:

Change medium and short date format for Slowenia like this:

Remove hardcoded dot in medium format

    "d. MMM. yyyy" → "d. MMM yyyy"

In order not to get "5. jan.. 2011" and remove
spaces in short format:

    "d. MM. yy"    → "d.M.yyyy"

Month names and day names updated to those from libicu48/CLDR to make
it the dots in the abbreviations fit with the new medium format.

AM/PM used instead of "dop." and "pop." because 24 hour mode is the
default for Slovenia, we almost always use AM/PM if 12 hour mode is
forced for countries where 24 hour mode is the default unless there is
a really strong reason not to.