Fixes: NB#288837 - Regional: inconsistencies in pakistan eng regional settings date format
RevBy: Berthold Krevert

Details:

It is weird that the order of day and month is different in the CLDR data
for en_PK to both en_IN *and* ur.

I can find no reason why English date formats in Pakistan should be
differently ordered from English date formats in India and English
date formats in the United Kingdom even if the order used
by ur.txt (Urdu in Pakistan) is also the same as the British order.

I.e. the order of day, month, and year should probably be the
same for all of these.

It looks like full date format for en_IN lacks the comma after EEEE,
I’ll add that as well.