Fixes: NB#288808 - Regional: inconsistencies in presentation date format of HK regional setting, language NOT Chinese
RevBy: Berthold Krevert

Details:

Swedish usually uses lower case month names.

If a language which uses lower case month and/or weekday names is
mixed into the pattern of a language which uses uppercase (here
English, Hongkong), then what to do?

Same problem if month or weekday names of language which has
grammatical inflections are mixed into the pattern of a different
language which does not have such inflections. The inflections would
look out of place there, especially if the month and/or weekday name
appears in a completely different position in the date format than in
the original context.

To reduce these problems of inflections and capitalizing, I use the
“stand-alone” forms of the month and weekday names always when such
mixing is done. These stand-alone forms lack grammatical inflections
and are “usually” uppercase, therefore they “usually” fit somewhat
better into a foreign date format.

But in Swedish, the stand-alone forms are lowercase in CLDR as well.

I think overall results will be better  if I change the stand-alone
versions of the Swedish month and weekday names to upper case as well.