Fixes: NB#290973 - Cannot share the default video “Dolby Soundscapes” when device language is Chinese.
RevBy: Holger Schröder, Jürg Billeter

Details:

Better don’t use [alternate shifted] in the Chinese collation tables
by default, MStringSearch sets it anyway.  If this is really needed in
MCollator as well, I should probably add a setter to the MCollator API
but it is better if this option is off by default in MCollator.