[libmlocale] Remove icu extradata. Contributes to JB#45113
The data hasn't been maintained for long time so it's possibly
doing more harm than good a this point. CLDR values could have been
updated further, and anyway the overrides have been sometimes
specific to Harmattan UI or for details that are not necessarily
used anywhere anymore.

In addition to that, the custom data here consumed over 5 megabytes
on a device.