[mlocale] Avoid broken MLocale::localeScripts() code and adjust tests. JB#52793
The locale script data seems removed from ICU in 2015 so
we need a reimplementation if there's any need for the method.
The method could be beneficial if there's wider use for the library
so made it just warns instead of total removal. But at the moment
the usage is quite thin in Sailfish side.

On non-ICU build this was defining the method but didn't have
an implementation. Made the method consistently defined now.

Disabled some related testing in ft_locales.

On testing side, also disabled @pinyinsearch collation which
didn't work and for which I didn't find any definitions.
Using that in real code would be making quite specific expectations
as it's not included in any documentation.

kk_Cyrl_KZ existence test removed. Deleted from ICU in 2015 CLDR 28
integration.